refactor(pages): inline queryPageByPath into findPageByPath
