+++ pages/src/index.ts
@@ -4,12 +4,7 @@ export { formatSlug } from './hooks/validateSlug.js'
 export { payloadPagesPlugin } from './plugin.js'
 export { findPageByPath } from './queries/findPageByPath.js'
 export { clearPathCache } from './queries/pathCache.js'
-export type {
-  FindPageByPathArgs,
-  PageDocument,
-  PageDocumentResult,
-  PagePathQueryArgs,
-} from './queries/types.js'
+export type { FindPageByPathArgs, PageDocument, PageDocumentResult } from './queries/types.js'
 export type { IncomingPageCollectionConfig as PageCollectionConfig } from './types/PageCollectionConfig.js'
 export type {
   IncomingPageCollectionConfigAttributes as PageCollectionIncomingConfigAttributes,
